width/height API change: size_t -> uint32_t.

+++ api.c
@@ -85,7 +85,7 @@ nvpipe_encode(nvpipe* const __restrict cdc,
               const size_t ibuf_sz,
               void* const __restrict obuf,
               size_t* const __restrict obuf_sz,
-              const size_t width, const size_t height, nvp_fmt_t format) {
+              const uint32_t width, const uint32_t height, nvp_fmt_t format) {
 	assert(cdc);
 	nvp_impl_t* enc = (nvp_impl_t*)cdc;
 	return enc->encode(enc, ibuf,ibuf_sz, obuf,obuf_sz, width,height, format);
@@ -96,8 +96,8 @@ nvpipe_decode(nvpipe* const __restrict codec,
               const void* const __restrict ibuf,
               const size_t ibuf_sz,
               void* const __restrict obuf,
-              size_t width,
-              size_t height) {
+              uint32_t width,
+              uint32_t height) {
 	assert(codec);
 	nvp_impl_t* dec = (nvp_impl_t*)codec;
 	return dec->decode(dec, ibuf,ibuf_sz, obuf, width,height);
